Move tests to the same directory as source code files
