+++ src/legacy/core_plugins/kbn_vislib_vis_types/public/components/options/metrics_axes/y_extents.test.tsx
@@ -19,9 +19,9 @@
 
 import React from 'react';
 import { mount, shallow } from 'enzyme';
-import { YExtents, YExtentsProps } from '../y_extents';
-import { ScaleTypes } from '../../../../utils/collections';
-import { NumberInputOption } from '../../../common';
+import { YExtents, YExtentsProps } from './y_extents';
+import { ScaleTypes } from '../../../utils/collections';
+import { NumberInputOption } from '../../common';
 
 describe('YExtents component', () => {
   let setMultipleValidity: jest.Mock;
